extract shared code from testing
